[BACKPORT 2.21.1][#22532] yugabyted-ui: Fix yugabyted UI Metrics is not getting displayed with k8s operator deployed clusters
Summary:
Original commit: 267720a1b36a62d185a922ff03a2aa0d0046b40d / D35322
Currently, the yugabyted UI is not displaying the metrics when deployed using the Kubernetes OSS operator.
This is because yugabyted API server gets the uuid of a node using the GetHostToUuidMap method which calls the 7000/tablet-servers endpoint and parses the HTML.
In the Kubernetes environment, the operator sets leader preferences, which adds some text to the page and breaks the HTML parsing.

This fix modifies the HTML parsing to account for this.

Test Plan: no test plan

Reviewers: nikhil

Reviewed By: nikhil

Subscribers: djiang, yugabyted-dev

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35368