[PLAT-13636] Add enhancements to replicated tables view
Summary:
**Context**
Prior to this diff, tables which were dropped from the YBDB on the source universe got omitted on the YBA UI.
Users would have to use yb-admin or YBA API to remove the table themselves as the UI didn't support showing these deleted tables.
The recommended and documented order of operations is:
1. Remove the table from replication on YBA
2. Drop the table on source, then drop table on target

The purpose of this change is to make it easier to clean up the dropped tables from the replication config
after perform in the table drop on the universe first accidentally.

Furthermore, this diff contains changes to allow sorting on the tables tab to improve the
user experience of finding tables.

**Change**
- Show tables which are dropped on the source universe
- Allow sorting by column
- Update delete table modal text
- Indicate the dropped tables in the select tables modal and inform the user that these tables will be dropped
  from the xCluster config after submitting the select table request.

Test Plan:
- Create an xCluster config and drop tables on the source universe.
- Verify that the user is able to see these dropped tables on the tables tab of the xCluster details page.
{F247980}
{F247981}
- Verify that the user is informed about these dropped tables on the select tables modal and these tables
  are not passed to the YBA backend when submitting the select table request.
{F179692}
- Verify that the user can now sort the tables by different columns.
{F247982}

Reviewers: rmadhavan, kkannan, cwang, hzare, lsangappa

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34971