[#22560] CDCSDK: Fix failure of test TestSlotRowDeletionWithMultipleStreams in tsan builds
Summary:
This diff fixes the failures of the test `CDCSDKConsumptionConsistentChangesTest.TestSlotRowDeletionWithMultipleStreams` in TSAN builds. The failures were because the earlier timeout of 10 seconds for the deletion of slot entry from the state table was insufficient. This diff increases the timeout by `kTimeMultiplier` (30 seconds) in case of sanitizer builds.
Jira: DB-11485

Test Plan: Jenkins: test regex: .*CDCSDKConsumptionConsistentChangesTest.TestSlotRowDeletion*

Reviewers: asrinivasan, skumar, siddharth.shah

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35367