[#21637] YSQL: Use role oid when storing session_authorization with YSQL Connection Manager
Summary:
In D34316, the provision was created to internally use role_oid instead of role as a GUC Variable when using Connection Manager. This patch replicates the same functionality for session_authorization (used to denote the session user), as both of these GUC variables are dependent directly on roles which can be ALTERed mid session. An example of this scenario:

```
SET session_authorization = a; (Connection 1)
ALTER ROLE a RENAME TO b; (Connection 2)
SHOW session_authorization; (Connection 1)
```

This would incorrectly show the session authorization as 'a' on Connection 1 when Connection Manager is enabled, as Connection Manager would cache the value of session_authorization as 'a' through the SET statement. Altering the role would not allow connection manager to change the cached value, as it could possibly happen on another connection as described in the scenario.

Jira: DB-10532

Test Plan: Jenkins: all tests

Reviewers: nkumar, mkumar

Reviewed By: mkumar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D35033