[PLAT-13939][PLAT-13925][PLAT-13943][PLAT-12172] Validate the task payload before triggering the audit_log_config
Summary:
This diff covers 4 small tickets:
1. [PLAT-13939] Validate the task payload before triggering the audit_log_config
```
If the existing audit log config on the universe is same as the requested audit log config in the modifyAuditLogConfig API, we now throw an error.
```
2. [PLAT-13925] Regex parsing error in the otel logs
```
When YCQL logs are under the category ERROR, the keyspace and scope are not present in the audit log. We now make these 2 optional.
```
3. [PLAT-13943] Audit logs are skipped for the logLevel type (INFO, NOTICE)
```
Default `log_min_messages` is "WARNING". Anything below that is not logged by default for YSQL. For API users, we leave the log level options as is and note down in API docs to set `log_min_messages` if log level below `WARNING` is required.
```
4. [PLAT-12172] Mark datadog logs with collector specific tags
```
We now add a couple of datadog specific labels `ddsource:yugabyte`, `service:yb-otel-collector`, `host:<nodeName>` to each audit log that gets exported to datadog. All these labels can always be overridden at the telemetry provider layer or the `modifyAuditLogging` API layer.
```

Test Plan:
Manually tested all 4 of the above scenarios.
Run UTs.
Run itests.

Reviewers: amalyshev

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35387