[PLAT-14112] Make ysql/ycql password as write-only and adjust its location
Summary:
Received feedback from @Kuriakose George Kulangara that it is not intuitive to see the ysql/ycql password property in the CreateSpec and rest of the properties like enable, enableAuth in a ysql object within the main Spec. These related properties should be found together.

This change is to move the password property into the main Spec along with other ysql and ycql properties and mark it writeOnly. This would show up in the documentation of that property. It is the responsibility of the YBA server to redact such writeOnly properties when returning it as part of a API result.

Test Plan: Unit test

Reviewers: dshubin, #yba-api-review!

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35335