[PLAT-13636] Remove dropped tables banner when there are no dropped tables
Summary:
This diff hides the dropped tables info banner when the dropped tables count is 0.
We hide this banner because the user doesn't gain much from it when the count is 0 and
we don't want to make the page unnecessarily crowded.

Test Plan:
Verify that the dropped tables info banner is hidden when the dropped tables count is 0.

Prior to this diff:
{F249291}
After applying this diff:
{F249290}

Reviewers: rmadhavan, kkannan

Reviewed By: kkannan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35452