[PLAT-14160]: Fix extraction of DB file for multi threads
Summary:
when two threds tries to read tar file at a time, their buffer could collide which corrupts the reads,
so made this block synchronized.
It is not expected to increase runtime as customer will probably add only one release and these files are supposed to be added only once.

Test Plan: Manually tested as this issue was easily reproducible through UI.

Reviewers: svarshney

Reviewed By: svarshney

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35455