[#22762] YSQL: Enable restarted webserver to serve requests in MiniCluster tests.
Summary:
5862233cc0847a73d13a8f4c4c26c9e129071e61 / D35116 added cleanup logic to the webserver on calling Stop().
This was implemented, in part, by using an atomic bool `stop_initiated` to indicate that the webserver was shutting down.

This revision resets this variable on calling Start(). This flow is invoked exclusively in Minicluster tests involving webserver restarts.
In production environments, webserver restarts involve terminating the background worker running the webserver and starting a new one,
and so the problem of restarts does not exist in such environments.

Jira: DB-11665

Test Plan:
```
./yb_build.sh --cxx-test integration-tests_cql-backup-test --gtest_filter CqlBackupTest.TestBackupRestoreWithoutTSWebUI -n 12
```

Reviewers: telgersma, loginov

Reviewed By: telgersma

Subscribers: ybase, smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35456