[PLAT-13806] Removes check that prevent rotating n2n certs alone
Summary:
rootAndClientRootCASame is a high level property that determines whether root CA for n2n & n2c
is same or not.
It could be the case that the cluster has only n2n enabled - where the previous check would fail given n2c is not configured.
Don't see the point for having that check in place as we already have dedicated flag for both n2n & n2c to determine whether to perform rotation or not.

Test Plan:
Created universe with n2n only.
Tried rotating n2n server certs.
Verified that the certificates are rotated.

Reviewers: sanketh, sneelakantan, #yba-api-review

Reviewed By: sneelakantan, #yba-api-review

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D34866