[#22615] YSQL: Incorrect seek key in prefix-based backward scan
Summary:
### Context

HybridScanChoices adds a kGroupEnd marker at the end of the DocKey. ScanChoices adds this marker at the end of the prefix in a prefix-based scan. This is acceptable in a forward scan since we wish to seek to the smallest key >= the scan target. However, in a backward scan, we wish to seek to the largest possible key <= the scan target. kGroupEnd marker is fighting against us in this case. We append kHighest (representing +Inf) to seek to the largest key but that happens after the kGroupEnd marker is appended. This is undesirable.

### Fix

Remove the kGroupEnd marker during the backward prefix-based seek scan (from a copy of current scan target).

### Impact

1. Uses a strict prefix-based scan.

Full key scans are exempt from this issue since the kGroupEnd is not added on a key column.

2. Is a backward scan.

As explained in the context section, kGroupEnd is fine when scanning forward since it is a small value.

3. Has IN clauses.

Without multiple options, we may not be changing the scan target based on the user provided input more than once. Seeks still happen, however, we do not append a kGroupEnd marker on new scan targets obtained directly from scanning the database.

### Scan Target Debugging

1. Scan Target in HybridScanChoices is not always followed by a kGroupEnd marker. This marker is used to end hash keys group and range keys group. When converting a DocKey to a human-readable string, we need this end of the keys marker. This change appends this marker so we do not get corruption errors when debugging the doc keys.

2. Propagate the allow_special flag when decoding keys so that special keys such as +Inf can be debug dumped appropriately.
Jira: DB-11524

Test Plan:
Jenkins

./yb_build.sh --java-test TestPgRegressDistinctPushdown

Added a test case that

1. Uses a prefix-based scan.
2. Is descending.
3. Has multiple Option Ranges (IN clause).

Backport-through: 2.18

Reviewers: tnayak, telgersma, smishra

Reviewed By: tnayak

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D35497