[BACKPORT 2024.1][#22624] xCluster: Fix issue where heartbeat metrics misses the needs_full_tablet_report
Summary:
Original commit: 5ee05f50a6a5f497763025e129abc4e14b3f6dc0 / D35467
Metrics in the tserver to master heartbeats are only sent every `tserver_heartbeat_metrics_interval_ms`. As part of this we also send the xCluster replication errors.
The errors are sent incrementally except when we sent the full report. The full report happens when there is a master leader failover. The previous heartbeat response has the request for a full report, and the next heartbeat may not include the metric report due to the interval. This causes us to miss sending the full xCluster error info to the new master.
This diff fixes this by recording the needs_full_tablet_report in PeriodicalHeartbeatDataProvider and using it whenever the next report is sent.

- Made the flag `tserver_heartbeat_metrics_interval_ms` runtime updatable.

Fixes #22624
Jira: DB-11531

Test Plan: XClusterTest.ReplicationErrorAfterMasterFailover

Reviewers: jhe, slingam, xCluster

Reviewed By: jhe

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35500