[#13358] YSQL: DDL Atomicity Stress Test
Summary:
This diff adds a new DDL atomicity stress test. The stress test setup a test
table with some data, then spawns 4 concurrent test threads. The first 3 of
these 4 threads continuously run a pair of complementing DDL statements for 10
iterations, the 4th thread runs a update DML statement for 10 iterations.

Thread 1: alter table add and drop column
Thread 2: add and drop columns with default values
Thread 3: create/drop an index on this table
Thread 4: thread continuously updates the rows on this table.

Because there are many concurrency between DDLs and DMLs, there will be execution
errors that are expected. The test lists all such expected errors. If an error
is one of these expected errors, the test will retry the failed statement until
eventually succeed. The test fails if an unexpected error is encountered.

When all the 4 threads have completed, the test does a number of verifications:

1. the table does not contain any additional columns
2. no indexes are present on this table
3. the rows on this table are updated correctly

The test adds some error injections that are controlled by new test gflags.

This diff also fixes a tsan data race that showed up in the newly added stress tests
by re-acquiring CatalogManager's mutex_ which is already released earlier in order
to fix a deadlock (commit ac184a8861f5cd4fbf7b6def02f7f107c676cbee).

Note that if we add more test scenarios with more concurrent DDL threads, we may
see more expected errors that are not already listed and the test would fail. In
that case we need to debug the failure. If it turns out that the error is
expected, we need to add the new expected error to the list of expected errors.
If the error indicates a bug, then we need to fix the bug to allow the test to
pass.

The stress test has two sets: non-colocated table and colocated table. A third
set for partitioned table was taken out of this diff. Partitioned table stress
test will be addressed by a separate bug fix diff.
Jira: DB-2996

Test Plan:
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityStressTest.BasicTest -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityStressTest.TestTxnVerificationFailure -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityStressTest.TestFailCatalogWrites -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityStressTest.TestFailDdlRollback -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityStressTest.TestFailDdlVerification -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityColocatedStressTest.BasicTest -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityColocatedStressTest.TestTxnVerificationFailure -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityColocatedStressTest.TestFailCatalogWrites -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityColocatedStressTest.TestFailDdlRollback -n 20 --tp 1
./yb_build.sh debug --sj  --cxx-test pg_ddl_atomicity_stress-test --gtest_filter=PgDdlAtomicityColocatedStressTest.TestFailDdlVerification -n 20 --tp 1

Reviewers: hsunder, zdrudi, fizaa

Reviewed By: zdrudi

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D32682