[PLAT-13452][Platform][Provider Validation]Err messages are getting reset if we hover/click away from the control
Summary:
Form errors are being reset on 'onBlur' event. This issue occurs due to the onBlur function in YBInput.tsx, which trims unwanted spaces at the end of the input. However, the input value is being set even when no trimming is necessary.

Fix:
A comparison has been added to check if the input value has trailing whitespace. The input value is now updated only when trimming is required.

Test Plan:
Tested manually
{F249715}

Reviewers: jmak

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35478