[#22571] DocDB: Use total number of bytes being flushed to understand whether writes should be stalled
Summary:
Currently we block write to RocksDB when there are 2 memtables are flushing.
It works well when RocksDB flushes are determined by the size of memtable.
But there are a lot of scenarios when we initiate RocksDB flush via some external events.
Such as snapshot, intents SST file deletion, schema change and so on.
So it is pretty common case when we could have 2 flushing memtables.

This diff introduces new behaviour to determine whether we should block writes.
The total number of bytes being flushed is used to determine whether writes should be blocked.

Also fixed bug that could appear when there are more than 2 flushing mem tables.
Suppose we have 3 memtables to flush.
Because of filters we could flush 2nd memtable, skipping 1st and 3rd.
Then 1st and 3rd memtables are flushed.
So we will get 2 sst files w/o continuous seqno range.
Fixed by stopping picking memtables when filter was not passed.
Jira: DB-11489

Test Plan: Jenkins

Reviewers: rthallam, arybochkin

Reviewed By: rthallam, arybochkin

Subscribers: yql, yugaware, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35204