[#22622] YSQL: Fix bitmap scan local recheck for columns that are not targets
Summary:
When an index produces results that require a recheck, bitmap scans recheck all the results at the Bitmap Table
scan node.

Bitmap scans can perform remote recheck and/or local recheck. Remote recheck is done by pushing down the
conditions.
Local recheck is done for conditions that cannot be pushed down.

If local recheck is done on a column that was not fetched from DocDB, it will fail, excluding the row from the
results. This produces incorrect results.

The solution is to add the columns referenced by Bitmap Scan's local quals to the DocDB request. Then the relevant
fields are present for the local qual evaluation.

**Why does this not affect normal index scans?** Normal index scans are simpler, because the request that has the original condition and the rechecking occur in the same node. Rechecks on Bitmap Scans are more complicated - a bitmap table scan may need to recheck a condition that was set up for the Bitmap Index scan only. There are more quals in more spots than the typical index scan.

=== Implementation Details ===
Since `ybcSetupTargets` is inside `ybcBeginScan` and both functions are used widely, we don't want to modify their
signature too much. The easiest way to communicate the new targets to `ybcSetupTargets` is to depend on a
datastructure that's already passed in: `Scan *pg_scan_plan`. If the `Scan *` node is a bitmap scan node, we can
extract the extra quals from it.

**How do we avoid always adding the extra quals?** Since the quals are added only if they exist, a simple way is
to remove the quals from the `Scan *` object if we don't need them. To avoid permanently modifying the plan
structure, a local copy is made.
Jira: DB-11530

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbBitmapScans'
```

Reviewers: tnayak, amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D35591