[#22630] DocDB: Don't set key bounds for intent iterator used during conflict resolution
Summary:
As part of conflict resolution, we create an intent iterator respecting the tablet's key bounds. This could lead to undetected conflicts and violation of serializable isolation guarantees.

For instance, consider a serializable transaction that did `select * from table`. It would lead to the following strong intent for empty doc key on all tablets of the table
```
SubDocKey(DocKey([], []), []) [kStrongRead] <ht> -> <txn>
```

During conflict resolution, an intent iterator is created using
```
      intent_iter_ = CreateIntentsIteratorWithHybridTimeFilter(
          doc_db_.intents, &status_manager(), doc_db_.key_bounds, &intent_key_upperbound_);
```

So essentially, when another transaction is trying to do an update, it requests weak read/write locks on the empty doc key `DocKey([], [])`. While checking conflicts for this record, we have the following
```
  Status ReadIntentConflicts(IntentTypeSet type, KeyBytes* intent_key_prefix) {
    ...
    // the below seeks to the tablet's lower key bound when the given target < tablet's lower key bound
    intent_iter_.Seek(intent_key_prefix->AsSlice());
    // and then we exit checking conflicts if the seeked intent record doesn't have prefix <intent_key_prefix>
```
which leads to undetected conflicts for empty doc key - `DocKey([], [])`. And the update/insert that should ideally have been blocked on the serializable txn instead goes through.

This issue is observed only when all of the following hold true
1. the tablet has key bounds set, which seems to happen only for tablets created through split.
2. there is a serializable transaction involved and other transaction are trying to perform a conflicting operation.

**Solution**
To address the problem, we create the intent iterator with `kNoBounds` thus allowing it to see all intents of the parent tablet. There should be no significant overhead for this approach, because as part of conflict resolution, the intent iterator seeks to the given entry (by the current transaction) and calls `Next` until the found key doesn't have the supplied key as a prefix. Since post split, the new requests at the child tablet correspond to the keys hosted by the child, it should not result in unnecessary additional `Seek`/`Next` calls. Additionally, compaction on the intentsdb doesn't use key bounds, so the intent record(s) on empty key are preserved through compactions, which is necessary for correctness.

If there is any perf impact, we should revisit this issue. Here are a couple of other relevant observations-
1. It is just the empty key that we had issues with in detecting conflicts. If serializable transaction are taken out of the picture, then we shouldn't have had hit such issues with just RC/RR transactions (since the conflict would have been at the key hash or higher, which falls within the tablet's key bounds).
2. If necessary, we can create the intent iterator with no lower bound, but the upper bound set to the tablet's upper bound.
Jira: DB-11537

Test Plan:
./yb_build.sh --cxx-test='TEST_F(PgTabletSplitTest, TestConflictResolutionChecksConflictsAgainstEmptyKey)'

The test fails consistently without the fix.

Reviewers: rthallam, timur, arybochkin, sergei

Reviewed By: arybochkin

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D35543