[PLAT-14250]: Do not promote autoflags on universe while deleting k8 universe
Summary:
Earlier, we used to promote autoflags while deleting k8s universe if it had xcluster, but this is not required after rollback.
We fixed it for VM universes earlier here: https://phorge.dev.yugabyte.com/D31469 but we forget to add it for k8s.

Test Plan: Tested manually by establishing a xcluster between k8s universe and verified that while deleting the universe autoflags are not promoted when both version are greater than rollback support version.

Reviewers: cwang, sanketh

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35630