[BACKPORT 2024.1][#21793] YSQL: Calculate Bitmap Scan recheck using initialized scan keys
Summary:
Original commit: a11e8a0b9131ba80771d9fabbfc3c197188baa4a / D35390
`YbPredetermineNeedsRecheck` was called with `*ybscan->keys, ybscan->nkeys`, which may not be fully initialised at the time of calling. The other use of `YbPredetermineNeedsRecheck` uses the keys from the `IndexScanState`. Doing the same thing for bitmap scans allows it to use the fully initialized scan keys.

To make it clear that it's now the caller's responsibility to determine if recheck is required, I included `recheck` as a parameter in the`yb_amgetbitmap_function`.
Jira: DB-10683

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbBitmapScans'
```

Reviewers: tnayak, amartsinchyk

Reviewed By: amartsinchyk

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35637