[#21813] YSQL: Make max_changes_in_memory flag configurable
Summary:
This diff introduces a GUC `yb_max_changes_in_memory` which specifies the maximum number of changes kept in memory per transaction, in reorder buffer. If the number of changes exceed this, then the access changes are spooled to the disk. The default value of this flag is 4096.

Before this change the `max_changes_in_memory` was a static constant with value 4096.
Jira: DB-10713

Test Plan: Jenkins: test regex: .*ReplicationSlot.*

Reviewers: stiwary, asrinivasan, skumar, siddharth.shah

Reviewed By: stiwary

Subscribers: yql, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35653