[BACKPORT 2.18][#22704] YSQL: Fix: IN expressions on single column ID 0 take the tuple path
Summary:
Original commit: 5aaf0c5d4bb1b458dae01d0f5008c475d01c0b93 / D35603
If you run the following query:
  SELECT * FROM pg_type WHERE typname IN ('regcollation','regconfig');

In the release build, it ends up taking the tuple path in ql_scanspec.cc:QLScanRange::Init() case QL_OP_IN, even though the WHERE clause is IN a single column (typname). The reason is that the index used in the plan, pg_type_typname_nsp_index, has typname as column 0, which is overloaded in the ql_scanspec.cc code to mean a tuple. This works ok in the release build because the DCHECK(options_itr->has_tuple_value()) is compiled out. The code seems to work anyway, but it's incorrect. In the debug build, kFirstColumnId is 10 instead of 0, so the check for ColumnId(0) in the code can never collide with a real column.

The fix is to instead use a nonempty ColumnValue.column_ids field to detect tuples. Also set the column id in the tuple case to kInvalidColumnId to avoid overloading any column IDs.
Jira: DB-11615

Test Plan:
Jenkins

Both of these tests trigger the DCHECK if ColumnId(0) is changed to kFirstColumnId in both places in ql_scanspec.cc before this change. They pass with this fix.
  ./yb_build.sh fastdebug --java-test org.yb.pgsql.TestPgSelect
  ./yb_build.sh fastdebug --java-test org.yb.pgsql.TestPgRegressHashInQueries

Reviewers: tnayak

Reviewed By: tnayak

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35668