[#22626] YSQL: Eagerly free the temporary heap tuple and its string representation.
Summary:
As part of streaming changes via logical replication, we create a temporary heap tuple. This is used to create the final heap tuple which is then streamed via the protocol to the client. This temporary heap tuple was being freed at the end of the transaction which was causing issues when a transaction consists of a large number of DMLs (10K+).

Additionally, we log the heap tuple being replicated via the logical replication for debugging purposes.

There were two problems with this:
1. Even though the logging was being done at `DEBUG2`, the string representation was being formed regardless of the log level. This string representation can be huge for large tuples, so it is costly.

2. We weren't freeing the string representation after logging. It was being done at the end of the transaction commit since this allocation is being done in the transaction memory context. As a result, for a large transaction, this memory would keep on accumulating and causes memory issues such as https://github.com/yugabyte/yugabyte-db/issues/22626.

This revision fixes the issues by eagerly freeing them right after use instead of relying on the transaction memory context cleanup. We now have the log statement inside an `if` statement. We also free the representation after logging.
Jira: DB-11533

Test Plan:
Jenkins: test regex: .*ReplicationSlot.*

Tested manually before and after the changes.

Before the change, the total memory allocated in the walsender at the start of ReorderBufferCommit: 136262064 bytes in 287 blocks; 13002296 free (367
chunks); 123259768 used

After the change, the above reduces to: Grand total: 27226544 bytes in 276 blocks; 16530312 free (361 chunks); 10696232 used. This is a ~11x reduction.

Reviewers: asrinivasan

Reviewed By: asrinivasan

Subscribers: yql, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D35663