[BACKPORT 2024.1][PLAT-14276][PLAT-14267] Fixing the openapi-format to 1.17.1 ver
Summary:
Original commit: e9fa07e6c3adee3f864b7e0c17bda7bca99cccde / D35690
As part of https://phorge.dev.yugabyte.com/D35315, I had fixed our builds to always use openapi-format@1.17.1 version. Always pulling and using the latest version of this tool is not good, as it might break our CI builds if there is a node version bump in the latest version of this tool. This was done in anticipation of such a scenario, but we unfortunately hit exactly this same scenario today. The change does ensure to use the exact version of 1.17.1 in a local build, but due to a bug, did not ensure that in the CI pipeline that uses npx.

This was exposed today as the upstream openapi-format tool released a new version (https://github.com/thim81/openapi-format/releases/tag/v1.18.0) that bumped the min nodejs version to 18.x (we use 12.22.11 and 16.10.0 on CI pipelines).

This change fixes the version of openapi-format tool used via npx command as well.

Test Plan: Phab pipeline

Reviewers: skurapati, dshubin

Reviewed By: skurapati

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35697