[#22825] DocDB: Vector Index General Read Path with DummyANN
Summary:
This change lays some foundations for the read path of vector indexes. It relies on a dummy implementation on the DocDB side that materializes all of the tablet's rows in memory before iterating through the closest rows to the query vector in increasing order of (distance from query vector, ybctid). This dummy in-memory implementation logic is added in `vectorann.cc` in the `DummyANN` class. This class implements an interface `VectorANN`, which later more sophisticated ANN index algorithms are expected to satisfy.

The `VectorANN` class expects to take in pairs of Vectors and Values. The Values are expected to be RocksDB keys that one can use to read RocksDB rows with. A `VectorANNIterator` can be instantiated with a query vector and is expected to yield rows stored in the VectorANN in increasing order of `(distance from query vector, ybctid)`. The `Iterator` interface has a `ANNPagingState` that can be used to page through this iterator. This paging state will have a pair of (distance from query vector) and (ybctid) to keep track of where we are within the iteration and to enable paged responses.

For each result from the vector search, DocDB needs to serialize the distance/score of this vector result from the query vector to allow for merging at the Pggate layer. For this reason, this diff adds a `distances` field to `PgsqlResponsePB`. The ith value of this contains the distance from the query vector to the ith response row. This merging logic on Pggate will be implemented in a follow-up diff. This is why this diff forces DummyANN vector index tables to have just one tablet.

Note that `DummyANN` is not MVCC aware. This is not a problem for this change as all visible rows are loaded into a `DummyANN` during read-time.

This new logic intends on grabbing a bunch of RocksDB keys from a `VectorANN` and doing point lookups in RocksDB. This is very much like what happened in `ExecuteBatchYbctids` before this change. In order to reuse that logic, this change abstracts the source of ybctids in `ExecuteBatchYbctids` by providing the method with an iterable `KeyProvider` class. This method was also renamed to `ExecuteBatchKeys` as it is possible in the future that Vector index keys might not just be a `ybctid`.

Other changes are:
- Removed the dependency of yb_vector library on yb_docdb -- the dependency should be the other way around.
- Renaming TestThreadHolder to ThreadHolder and moving it to the util library, since it is used for parallelizing index load and validation in hnsw_tool.

**Upgrade/Rollback safety:**
This adds vector index protobuf fields that should not be used by anybody production customer right now.

Jira: DB-11724

Test Plan: Jenkins: test regex:  .*TestPgRegressThirdPartyExtensionsPgvector.*

Reviewers: sergei, mbautin

Reviewed By: sergei, mbautin

Subscribers: svc_phabricator, robert, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D35708