[#18611, #18616] YSQL: Throw correct error string and code in case of a deadlock and enable query layer retries for deadlock errors
Summary:
When a transaction is aborted due to a deadlock the correct error to be thrown
to the external user is "deadlock detected" with the error code 40P01. However,
sometimes YSQL doesn't throw the correct error string because of the following
reason: `GetStatusMsgAndArgumentsByCode()` in YSQL doesn't always see
pg_err_code to be ERRCODE_T_R_DEADLOCK_DETECTED.

pg_err_code is calculated in YSQL in FetchErrorCode(). If PgClientSession
returns a status with a specific pg_err_code, that is used. Otherwise, it is
derived based on the txn_err_code.

When a deadlock occurs and a transaction is aborted, different errors can occur
on different transaction participant nodes, an example is below:

Error from node1:

```Operation expired (yb/docdb/deadlock_detector.cc:637): Transaction 03a4718b-6f02-4226-bbb2-0d1898687693 aborted due to a deadlock.
03a4718b-6f02-4226-bbb2-0d1898687693 -> f888235c-165f-4d8e-ac58-cfb48e47a0bb -> 03a4718b-6f02-4226-bbb2-0d1898687693 : kDeadlock (pgsql error 40P01) (transaction error 6),```

Error from node2:

```Operation failed. Try again (yb/tablet/transaction_participant.cc:290): Transaction was recently aborted: 03a4718b-6f02-4226-bbb2-0d1898687693: 40001 (pgsql error 40001) (transaction error 1)```

In the above situation, the deadlock error from node1 doesn't have pg_err_code
set. And the second one has a pg_err_code of 40001. This results in the final
error being 40001 instead of 40P01. This is fixed by setting pg_err_code to
40P01 in `GetTransactionDeadlockStatus()`.

Even after setting it to 40P01, YSQL receives no pg_err_code because
PgClientSession receives different pg_err_codes from different nodes (see AppendPsqlErrorCode()).
In this case, YSQL relies on the txn_err_code to derive the pg_err_code (as per FetchErrorCode()).
So, we need to ensure that the txn_err_code is kDeadlock even if there are kAbort errors from other
nodes. This is ensured in AppendTxnErrorCode() as part of this diff.

Enabling query layer retries for kDeadlock errors:
-------------------------------------------------
This diff also adds query later retries for kDeadlock (and kAborted) errors whenever possible. The following important changes were made to enable this:
(1) Retrying kAborted and kDeadlock errors is done by restarting the whole transaction (for RC transactions too). For RC isolation, this is unlike retries from kConflict and kReadRestart errors which can be retried by redoing just the statement facing the error. kDeadlock and kAborted error retries can't undo and redo the statement because the transaction is already aborted. So, retries are done by restarting the whole transaction. Also, they are only possible if no data has been sent to the client yet for the transaction.
(2) The logic to restart a transaction has been modified in postgres.c. This was done because of the following reasoning:

a. Earlier, all query layer retries for a certain statement would either go via the `yb_restart_current_stmt` (or) `yb_restart_transaction` for all retries, and it was never a mix. But now, for RC, we can have some retries do `yb_restart_current_stmt` (e.g., for kConflict/kReadRestart in a txn block) and some retries can do `yb_restart_transaction` (e.g., for kDeadlock/kAborted errors).

b. A problem can occur if we just enable kDeadlock/Aborted retries to use `yb_restart_transaction` without any changes. This logic only restarts the YB-side transaction (the PG side transaction state is reused). If we retry a kDeadlock/kAborted error for the first statement in a transaction block (i.e., inside `BEGIN;`), the restart will not result in PG again taking an internal savepoint which is required by all statements in RC txn blocks. This is because PG acquires an internal savepoint for each RC statement once in StartTransactionCommand(), and we don't redo the PG transaction's state machine transition again during this transaction restart logic. If the query then faces a kConflict error, it would go into `yb_restart_current_stmt` and fail because there was no internal savepoint to undo the statement's partial work. This is solved by checking if we are in a transaction block in `yb_restart_transaction`, and if we are, explicitly acquiring of an internal savepoint for the statement before re-executing the statement.

Test Plan:
Jenkins

Backport-through: 2024.1

Reviewers: bkolagani, tvesely, patnaik.balivada

Reviewed By: bkolagani, patnaik.balivada

Subscribers: patnaik.balivada, ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32251